refactor(workflow): lift compute_next_run into wardian-core::schedule
